图片处理/weiImage/requirements.txt
aspy.yaml==1.3.0
certifi==2019.9.11
cfgv==2.0.1
chardet==3.0.4
Click==7.0
colorama==0.4.0
cycler==0.10.0
filelock==3.0.12
Flask==1.1.1
future==0.18.2
identify==1.4.7
idna==2.8
importlib-metadata==0.23
importlib-resources==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==7.2.0
nodeenv==1.3.3
packaging==19.2
Pillow==5.3.0
pluggy==0.13.0
pre-commit==1.20.0
prettytable==0.7.2
py==1.8.0
pyparsing==2.4.5
pypng==0.0.20
PyQRCode==1.2.1
pytz==2018.7
PyYAML==5.1.2
qrcode==6.0
requests==2.22.0
six==1.11.0
toml==0.10.0
tox==3.14.1
translate==3.5.0
urllib3==1.25.6
virtualenv==16.7.7
Werkzeug==0.16.0

